+++ lib/dart_otp.dart
@@ -1,0 +1,13 @@
+///
+/// @project  : dart_otp
+/// @author   : Gin (gin.lance.inside@hotmail.com)
+/// @link     : https://github.com/BrunoMiguens/dart_otp
+/// @Disc     : a dart package to generate and verify one-time passwords
+///
+library dart_otp;
+
+export 'src/hotp.dart';
+export 'src/totp.dart';
+
+export 'src/components/otp_algorithm.dart';
+export 'src/components/otp_type.dart';