+++ lib/src/hotp.dart
@@ -4,70 +4,74 @@
 ///
 
 import 'otp.dart';
+import 'package:dart_otp/src/components/otp_type.dart';
+import 'package:dart_otp/src/components/otp_algorithm.dart';
 
+///
+/// HOTP class will generate the OTP (One Time Password) object with given counter.
 class HOTP extends OTP {
-  HOTP(String secret, [int digits = 6]) : super(secret, digits);
+  /// The counter value for one-time password.
+  int counter;
 
-  ///
-  /// Generate the OTP with the given count
-  ///
-  /// @param {count}
-  /// @type {int}
-  /// @desc the OTP HMAC counter
-  ///
-  /// @return {String}
-  ///
-  /// @example
-  /// HOTP hotp = dotp.HOTP('BASE32_ENCODED_SECRET');
-  /// hotp.at(0); // => 432143
-  ///
-  String at(int count) {
-    return super.generateOTP(count);
-  }
-
-  ///
-  /// Verifies the OTP passed in against the current time OTP.
-  ///
-  /// @param {otp}
-  /// @type {String}
-  /// @desc the OTP waiting for checking
-  ///
-  /// @param {counter}
-  /// @type {int}
-  /// @desc the OTP HMAC counter
-  ///
-  /// @return {Boolean}
-  ///
-  /// @example
-  /// TOTP totp = dotp.TOTP('BASE32ENCODEDSECRET');
-  /// totp.now(); // => 432143
-  /// // Verify for current time
-  /// totp.verify(432143); // => true
-  /// // Verify after 30s
-  /// totp.verify(432143); // => false
-  ///
-  bool verify(String otp, int counter) {
-    String otpCount = this.at(counter);
-
-    if (otp == otpCount) {
-      return true;
-    }
-    return false;
-  }
-
-  ///
-  /// Generate a url with TOTP instance.
-  ///
-  /// @param {issuer}
-  /// @type {String}
-  /// @desc maybe it is the Service name
-  ///
-  /// @return {String}
-  ///
   @override
-  String urlGen(String _issuer, String _type) {
-    _issuer ??= '';
-    _type = 'totp';
-    return super.urlGen(_issuer, _type);
+  OTPType get type => OTPType.HOTP;
+
+  @override
+  Map<String, dynamic> get extraUrlProperties => {"counter": counter};
+
+  ///
+  /// This constructor will create an HOTP instance.
+  ///
+  /// All parameters are mandatory however [counter],
+  /// [digits] and [algorithm] have a default values, so can be ignored.
+  ///
+  /// Will throw an exception if the line above isn't satisfied.
+  ///
+  HOTP(
+      {String secret,
+      int counter = 0,
+      int digits = 6,
+      OTPAlgorithm algorithm = OTPAlgorithm.SHA1})
+      : super(secret: secret, digits: digits, algorithm: algorithm) {
+    this.counter = counter;
+  }
+
+  ///
+  /// Generate the HOTP value with the given count
+  ///
+  /// ```dart
+  /// HOTP hotp = HOTP(secret: 'BASE32ENCODEDSECRET');
+  /// hotp.at(counter: 0); // => 432143
+  /// ```
+  ///
+  String at({int counter}) {
+    if (counter == null || counter < 0) {
+      return null;
+    }
+
+    return super.generateOTP(input: counter);
+  }
+
+  ///
+  /// Verifies the HOTP value passed in against the a given counter.
+  ///
+  /// All parameters are mandatory.
+  ///
+  /// ```dart
+  /// HOTP hotp = HOTP(secret: 'BASE32ENCODEDSECRET');
+  /// hotp.at(counter: 0); // => 432143
+  /// // Verify for current time
+  /// hotp.verify(otp: 432143, counter: 0); // => true
+  /// // Verify after 30s
+  /// hotp.verify(otp: 432143, counter: 10); // => false
+  /// ```
+  ///
+  bool verify({String otp, int counter}) {
+    if (otp == null || counter == null) {
+      return false;
+    }
+
+    String otpCount = this.at(counter: counter);
+    return otp == otpCount;
   }
 }