+++ lib/src/otp.dart
@@ -4,34 +4,47 @@
 ///
 
 import 'dart:math';
-import 'package:crypto/crypto.dart';
 import 'package:base32/base32.dart';
-import 'util.dart';
+import 'package:dart_otp/src/components/otp_algorithm.dart';
+import 'package:dart_otp/src/components/otp_type.dart';
+import 'package:dart_otp/src/utils/generic_util.dart';
+import 'package:dart_otp/src/utils/algorithm_util.dart';
+import 'package:dart_otp/src/utils/otp_util.dart';
 
 abstract class OTP {
-  ///
-  /// This constructor will create OTP instance.
-  ///
-  /// @param {secret}
-  /// @type {String}
-  /// @desc random base32-encoded key, it is the
-  /// key that be used to verify.
-  ///
-  /// @param {digits}
-  /// @type {int}
-  /// @desc the length of the one-time password, default to be 6
-  ///
-  /// @param {digest}
-  /// @type {String}
-  /// @desc the key that be used to do HMAC encoding, dedault and
-  /// only to be "sha1"
-  ///
-  ///
-  String secret;
+  /// The length of the one-time password, between 6 and 8.
   int digits;
-  OTP(String secret, [int digits = 6]) {
+
+  /// The Base32 secret key used to generate the one-time password.
+  String secret;
+
+  /// The crypto algorithm used on HMAC encoding.
+  OTPAlgorithm algorithm;
+
+  /// The type of the token.
+  OTPType get type;
+
+  /// To access custom properties when generating the url.
+  Map<String, dynamic> get extraUrlProperties;
+
+  ///
+  /// This constructor will create an OTP instance.
+  ///
+  /// All parameters are mandatory however [digits] and
+  /// [algorithm] have a default value, so can be ignored.
+  /// Will throw an exception if the line above isn't satisfied.
+  ///
+  OTP(
+      {String secret,
+      int digits = 6,
+      OTPAlgorithm algorithm = OTPAlgorithm.SHA1})
+      : assert(secret != null),
+        assert(digits != null),
+        assert(algorithm != null),
+        assert(digits >= 6 && digits <= 8) {
     this.secret = secret;
     this.digits = digits;
+    this.algorithm = algorithm;
   }
 
   ///
@@ -39,22 +52,19 @@ abstract class OTP {
   /// function, it will generate the OTP object with params,
   /// the params may be counter or time.
   ///
-  /// @param {input}
-  /// @type {int}
-  /// @desc input params to generate OTP object, maybe
-  /// counter or time.
+  /// All parameters are mandatory however [algorithm] have
+  /// a default value, so can be ignored.
   ///
-  /// @return {String}
-  ///
-  String generateOTP(int input) {
+  String generateOTP({int input, OTPAlgorithm algorithm = OTPAlgorithm.SHA1}) {
     /// base32 decode the secret
     var hmacKey = base32.decode(this.secret);
 
     /// initial the HMAC-SHA1 object
-    var hmacSha1 = Hmac(sha1, hmacKey);
+    var hmacSha =
+        AlgorithmUtil.createHmacFor(algorithm: algorithm, key: hmacKey);
 
     /// get hmac answer
-    var hmac = hmacSha1.convert(Util.intToBytelist(input)).bytes;
+    var hmac = hmacSha.convert(Util.intToBytelist(input: input)).bytes;
 
     /// calculate the init offset
     int offset = hmac[hmac.length - 1] & 0xf;
@@ -73,20 +83,23 @@ abstract class OTP {
   }
 
   ///
-  /// Generate a url with TOTP or HOTP instance.
+  /// Generate a url with OTP instance.
   ///
-  /// @param {issuer}
-  /// @type {String}
-  /// @desc maybe it is the Service name
+  /// Use [issuer] and [account] parameters to specify the token information.
+  /// All the remaining OTP fields will be exported.
   ///
-  /// @param {type}
-  /// @type {String}
-  /// @desc type of OTP instance
-  ///
-  /// @return {String}
-  ///
-  String urlGen(String _issuer, String _type) {
+  String generateUrl({String issuer, String account}) {
     final _secret = this.secret;
-    return 'otpauth://$_type/SK?secret=$_secret&issuer=$_issuer';
+    final _type = OTPUtil.otpTypeValue(type: type);
+    final _account = Uri.encodeComponent(account ?? '');
+    final _issuer = Uri.encodeQueryComponent(issuer ?? '');
+
+    final _algorithm = AlgorithmUtil.rawValue(algorithm: algorithm);
+    final _extra = extraUrlProperties
+        .map((key, value) => MapEntry(key, "$key=$value"))
+        .values
+        .join('&');
+
+    return 'otpauth://$_type/$_account?secret=$_secret&issuer=$_issuer&digits=$digits&algorithm=$_algorithm&$_extra';
   }
 }