+++ CHANGELOG.md
@@ -1,9 +1,43 @@
 ## [1.0.1]
 
-* release the first stable version
+* Release the first stable version
 
 ## [1.0.2]
 
-* add example file
-* add docs
-* fix dart format warning
+* Add docs
+* Add example file
+* Fix dart format warning
+
+## [1.0.3]
+
+* Add documentation
+* Minor code style enhancements
+* Make interval and digits properties public for time based tokens
+
+## [1.1.0]
+
+* Add unit tests
+* Update documentation
+* Update components API
+* Refactor some components
+* Minor code style enhancements
+* Increase options export option for generate url function
+
+## [1.2.0]
+
+* Minor enhancements
+* Add new counter property inside hotp object
+* Update documentation according to dart guidelines
+* Add support for SHA256 algorithm (available: SHA1 and SHA256)
+* Add extra url proprties to otp object in order to export:
+  * Digits
+  * Issuer
+  * Period
+  * Account
+  * Counter
+  * Algorithm
+
+## [1.3.0]
+
+* Minor enhancements
+* Add support for SHA384 and SHA512 algorithm (available: SHA1, SHA256, SHA384 and SHA512)